0.7.0 release (#2701)
* 0.7.0 release

* nbstripout

* ViT-Huge is too Huge?

* Try disabling pip install torchgeo in notebook

* Fix SemanticSegmentationTask tests

* ViT-Large is too Large?

* No persistent storage between tests

* Add quotes

* No ViT tests

* Revert "nbstripout"

This reverts commit a6e45eae1c3b31c67f2d8a2d7b75a6074dec7daa.

* Remove unrecognized property

* Fix issues spotted in docs

* Fix bugs in xres/yres

* Undo change to notebook

* Increase coverage

* Renable tests

* Remove unused parameter resulting in 3x repeated tests

* Use local dir instead of /tmp

* Log storage and separate components

* Only keep a single tmp_path at a time

* Only test vit

* Only test samplers and trainers

---------

+++ torchgeo/samplers/utils.py
@@ -29,7 +29,7 @@ def _to_tuple(value: tuple[float, float] | float) -> tuple[float, float]:
     Returns:
         value if value is a tuple, else (value, value)
     """
-    if isinstance(value, float | int):
+    if isinstance(value, int | float):
         return (value, value)
     else:
         return value
